test/topology: move topology helpers to common file
Move helper functions to a common file ahead of splitting topology
tests.
