+++ test/topology/test_topology.py
@@ -20,71 +20,11 @@ from test.pylib.manager_client import ManagerClient
 from cassandra.cluster import Session
 from test.pylib.random_tables import RandomTables
 from test.pylib.rest_client import inject_error_one_shot
+from test.topology.util import get_token_ring_host_ids, get_current_group0_config, \
+                               check_token_ring_and_group0_consistency,            \
+                               wait_for_token_ring_and_group0_consistency
 
 logger = logging.getLogger(__name__)
-
-
-async def get_token_ring_host_ids(manager: ManagerClient, srv: ServerInfo) -> set[str]:
-    """Get the host IDs of normal token owners known by `srv`."""
-    token_endpoint_map = await manager.api.client.get_json("/storage_service/tokens_endpoint", srv.ip_addr)
-    normal_endpoints = {e["value"] for e in token_endpoint_map}
-    logger.info(f"Normal endpoints' IPs by {srv}: {normal_endpoints}")
-    host_id_map = await manager.api.client.get_json('/storage_service/host_id', srv.ip_addr)
-    all_host_ids = {e["value"] for e in host_id_map}
-    logger.info(f"All host IDs by {srv}: {all_host_ids}")
-    normal_host_ids = {e["value"] for e in host_id_map if e["key"] in normal_endpoints}
-    logger.info(f"Normal endpoints' host IDs by {srv}: {normal_host_ids}")
-    return normal_host_ids
-
-
-async def get_current_group0_config(manager: ManagerClient, srv: ServerInfo) -> set[tuple[str, bool]]:
-    """Get the current Raft group 0 configuration known by `srv`.
-       The first element of each tuple is the Raft ID of the node (which is equal to the Host ID),
-       the second element indicates whether the node is a voter.
-     """
-    assert(manager.cql)
-    host = (await wait_for_cql_and_get_hosts(manager.cql, [srv], time.time() + 60))[0]
-    group0_id = (await manager.cql.run_async(
-        "select value from system.scylla_local where key = 'raft_group0_id'",
-        host=host))[0].value
-    config = await manager.cql.run_async(
-        f"select server_id, can_vote from system.raft_state where group_id = {group0_id} and disposition = 'CURRENT'",
-        host=host)
-    result = {(str(m.server_id), bool(m.can_vote)) for m in config}
-    logger.info(f"Group 0 members by {srv}: {result}")
-    return result
-
-
-async def check_token_ring_and_group0_consistency(manager: ManagerClient) -> None:
-    """Ensure that the normal token owners and group 0 members match
-       according to each currently running server.
-    """
-    servers = await manager.running_servers()
-    for srv in servers:
-        group0_members = await get_current_group0_config(manager, srv)
-        group0_ids = {m[0] for m in group0_members}
-        token_ring_ids = await get_token_ring_host_ids(manager, srv)
-        assert token_ring_ids == group0_ids
-
-
-async def wait_for_token_ring_and_group0_consistency(manager: ManagerClient, deadline: float) -> None:
-    """Weaker version of the above check; the token ring is not immediately updated after
-    bootstrap/replace/decommission - the normal tokens of the new node propagate through gossip.
-    Take this into account and wait for the equality condition to hold, with a timeout.
-    """
-    servers = await manager.running_servers()
-    for srv in servers:
-        group0_members = await get_current_group0_config(manager, srv)
-        group0_ids = {m[0] for m in group0_members}
-        async def token_ring_matches():
-            token_ring_ids = await get_token_ring_host_ids(manager, srv)
-            diff = token_ring_ids ^ group0_ids
-            if diff:
-                logger.warning(f"Group 0 members and token ring members don't yet match" \
-                               f" according to {srv}, symmetric difference: {diff}")
-                return None
-            return True
-        await wait_for(token_ring_matches, deadline, period=.5)
 
 
 @pytest.mark.asyncio
